--- hds-notary/library/src/main/java/pt/ulisboa/tecnico/sec/client/library/exceptions/GoodWrongOwner.java
@@ -1,24 +1,0 @@
-package pt.ulisboa.tecnico.sec.client.library.exceptions;
-
-public class GoodWrongOwner extends ServerException {
-
-    public GoodWrongOwner() {
-    }
-
-    public GoodWrongOwner(String message) {
-        super(message);
-    }
-
-    public GoodWrongOwner(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public GoodWrongOwner(Throwable cause) {
-        super(cause);
-    }
-
-    public GoodWrongOwner(String message, Throwable cause, boolean enableSuppression,
-        boolean writableStackTrace) {
-        super(message, cause, enableSuppression, writableStackTrace);
-    }
-}
