--- hds-notary/library/src/main/java/pt/ulisboa/tecnico/sec/client/library/exceptions/UserNotFoundException.java
@@ -1,26 +1,0 @@
-package pt.ulisboa.tecnico.sec.client.library.exceptions;
-
-public class UserNotFoundException extends ServerException {
-
-    public UserNotFoundException() {
-    }
-
-    public UserNotFoundException(String message) {
-        super(message);
-    }
-
-    public UserNotFoundException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public UserNotFoundException(Throwable cause) {
-        super(cause);
-    }
-
-    public UserNotFoundException(String message,
-        Throwable cause,
-        boolean enableSuppression,
-        boolean writableStackTrace) {
-        super(message, cause, enableSuppression, writableStackTrace);
-    }
-}
