+++ hds-notary/library/src/main/java/pt/ulisboa/tecnico/sec/library/exceptions/GoodNotFoundException.java
@@ -1,0 +1,26 @@
+package pt.ulisboa.tecnico.sec.library.exceptions;
+
+public class GoodNotFoundException extends ServerException {
+
+    public GoodNotFoundException() {
+    }
+
+    public GoodNotFoundException(String message) {
+        super(message);
+    }
+
+    public GoodNotFoundException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public GoodNotFoundException(Throwable cause) {
+        super(cause);
+    }
+
+    public GoodNotFoundException(String message,
+        Throwable cause,
+        boolean enableSuppression,
+        boolean writableStackTrace) {
+        super(message, cause, enableSuppression, writableStackTrace);
+    }
+}
