+++ hds-notary/library/src/main/java/pt/ulisboa/tecnico/sec/library/exceptions/ServerException.java
@@ -1,0 +1,24 @@
+package pt.ulisboa.tecnico.sec.library.exceptions;
+
+public class ServerException extends Exception {
+
+    public ServerException() {
+    }
+
+    public ServerException(String message) {
+        super(message);
+    }
+
+    public ServerException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public ServerException(Throwable cause) {
+        super(cause);
+    }
+
+    public ServerException(String message, Throwable cause, boolean enableSuppression,
+        boolean writableStackTrace) {
+        super(message, cause, enableSuppression, writableStackTrace);
+    }
+}
