+++ hds-notary/server/pom.xml
@@ -13,7 +13,7 @@
                     <configuration>
                         <!-- to prevent deprecation warning: -->
                         <killAfter>-1</killAfter>
-                        <mainClass>pt.ulisboa.tecnico.sec.client.server.HdsNotaryApplication
+                        <mainClass>pt.ulisboa.tecnico.sec.server.HdsNotaryApplication
                         </mainClass>
                         <systemProperties>
                         </systemProperties>
